refactor(search-query): remove role search query and add search query to common class
