+++ domain/src/test/java/com/kaua/ecommerce/users/domain/pagination/SearchQueryTest.java
@@ -1,19 +1,19 @@
-package com.kaua.ecommerce.users.domain.roles;
+package com.kaua.ecommerce.users.domain.pagination;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class RoleSearchQueryTest {
+public class SearchQueryTest {
 
     @Test
-    void givenAValidValues_whenCallNewRoleSearchQuery_shouldReturnARoleSearchQueryInstancie() {
+    void givenAValidValues_whenCallNewSearchQuery_shouldReturnASearchQueryInstancie() {
         final var page = 0;
         final var perPage = 10;
         final var terms = "";
         final var sort = "createdAt";
         final var direction = "asc";
 
-        final var aQuery = new RoleSearchQuery(page, perPage, terms, sort, direction);
+        final var aQuery = new SearchQuery(page, perPage, terms, sort, direction);
 
         Assertions.assertEquals(page, aQuery.page());
         Assertions.assertEquals(perPage, aQuery.perPage());
