+++ infrastructure/src/main/java/com/kaua/ecommerce/users/infrastructure/api/controllers/RoleController.java
@@ -10,7 +10,7 @@ import com.kaua.ecommerce.users.application.role.retrieve.list.ListRolesUseCase;
 import com.kaua.ecommerce.users.application.role.update.UpdateRoleCommand;
 import com.kaua.ecommerce.users.application.role.update.UpdateRoleUseCase;
 import com.kaua.ecommerce.users.domain.pagination.Pagination;
-import com.kaua.ecommerce.users.domain.roles.RoleSearchQuery;
+import com.kaua.ecommerce.users.domain.pagination.SearchQuery;
 import com.kaua.ecommerce.users.infrastructure.api.RoleAPI;
 import com.kaua.ecommerce.users.infrastructure.roles.models.CreateRoleApiInput;
 import com.kaua.ecommerce.users.infrastructure.roles.models.GetRoleResponse;
@@ -86,7 +86,7 @@ public class RoleController implements RoleAPI {
     @Override
     public Pagination<ListRoleResponse> listRoles(String search, int page, int perPage, String sort, String direction) {
         return this.listRolesUseCase
-                .execute(new RoleSearchQuery(page, perPage, search, sort, direction))
+                .execute(new SearchQuery(page, perPage, search, sort, direction))
                 .map(RoleApiPresenter::present);
     }
